Fix answer totals displaying incorrectly
Survey answers was displaying incorrect totals as the `response.answers`
was coming back with all answers across every survey. This meant that
if the answer matched an answer from another survey exactly,
case-sensitivity included, that it was included in another total.
A `.filter()` was placed on the array to look for the clicked on surveys
id denoted as surveyRef on the model, and `store.survey._id`.